requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
cssselect==1.1.0
cssutils==2.5.0
defusedxml==0.7.1
Django==3.2.14
django-allauth==0.51.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-filter==22.1
django-phonenumber-field==6.3.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-twilio==0.13.2
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
flake8==4.0.1
idna==3.3
install==1.3.5
lxml==4.9.1
Markdown==3.3.7
mccabe==0.6.1
oauthlib==3.2.0
phonenumbers==8.12.51
Pillow==9.2.0
premailer==3.10.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
twilio==6.63.2
urllib3==1.26.10
yagmail==0.15.280
